Fix build.
* UIProcess/API/Cocoa/WKProcessClassConfiguration.mm:
(-[WKProcessClassConfiguration description]):

Canonical link: https://commits.webkit.org/147016@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164306 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/UIProcess/API/Cocoa/WKProcessClassConfiguration.mm
@@ -46,7 +46,7 @@
 
 - (NSString *)description
 {
-    NSString *description = [NSString stringWithFormat:@"<%@: %p; maximumProcessCount = %ld", NSStringFromClass(self.class), self, _maximumProcessCount];
+    NSString *description = [NSString stringWithFormat:@"<%@: %p; maximumProcessCount = %lu", NSStringFromClass(self.class), self, static_cast<unsigned long>(_maximumProcessCount)];
     if (_injectedBundleURL)
         return [description stringByAppendingFormat:@"; injectedBundleURL: \"%@\">", _injectedBundleURL.get()];
